all: update the go.mod
Set the minimum Go to 1.18 and update the share modules.

When this program written, the ini library does not have the Unmarshal
function, so we load the configuration by reading each key and parse it
manually.

Now that we have Unmarshal function, the way we parse the configuration
is simplified so does the way configuration written.
This changes the accept_backend, capture_request_header written.
Instead of using comma to set multiple values, now it must be written
one key and one value on different line.