all: restructure the log that we send to influxdb
Since the influxdb v2, and since the flux query language introduced,
the way the log read and queried kinda changes.

Things that we can query on v1, is not possible (or maybe I forgotten)
anymore.

This changes move all haproxy log fields value that is not number (except
HTTP status code) to tags and left all numbers (like time, number of
connections, bytes read) in the fields.