Release haminer v0.3.0 (2025-12-29)
**💧 Group documentation into _doc/ directory**.

This is so we can serve the documentation under
https://kilabit.info/project/haminer.

In the Makefile, we add task `serve-doc` to preview the documentation.

**💧 Replace module "share" with "pakakeh.go"**

The "share" module has been renamed to "pakakeh.go" and moved to different
git repository.

**💧 Realign struct HttpLog and questdbClient**

The HttpLog realign from 304 to 288 bytes.
The questdbClient realign from 56 to 24 bytes.

**🪵 Comply with all linters recommendations**

Some of breaking changes,

* Field [Config.HttpUrl] renamed to [Config.HTTPURL]
* Field [ConfigForwarder.Url] renamed to [ConfigForwarder.URL]
* Struct [HttpLog] renamed to [HTTPLog]

**💧 _ops/haminer-test: setup container for development using mkosi**

The container run HAProxy and PostgreSQL servers.

We also create new dummy backend to test backend in HAProxy.
The haminer-dummy-backend run in container and serve two ports in HTTP
and TCP modes.

**🌱 all: implement forwarder for Postgresql**

The Postgresql forwarder accept single option "URL",

  [forwarder "postgresql"]
  url = postgres://<user>:<pass>@<host>/<database>?sslmode=<>

The user and database must already created first, manually.

**🌼 Do not log ServerName with invalid connection '<NOSRV>'**

**🌼 _AUR: fix installation of binary in package function**

**Add linter gocheck**

Program gocheck implement go static analysis using [Analyzer] that are not
included in the default go vet.
See package [lib/goanalysis] for more information.

[Analyzer]: https://pkg.go.dev/golang.org/x/tools/go/analysis#hdr-Analyzer +
[lib/goanalysis]: https://pkg.go.dev/git.sr.ht/~shulhan/pakakeh.go/lib/goanalysis/